// types/server/contracts.d.ts
  interface QueryRes<T> {
    query_result: T;
  }
// ...
  type Holding = { address: string; amount: string };
  interface Holdings {
    locked_native: Holding[];
    locked_cw20: Holding[];
    liquid_native: Holding[];
    liquid_cw20: Holding[];
    is_placeholder?: true;
  }
// ...
  type ClaimInquiry = {
    is_claimed: boolean;
  };
// ...
  type VaultRateInfo = {
    vault_addr: string; //"terra172u..
    fx_rate: string; //"1.206784043460040765"
  };